Speed-up and fix bug in condition evaluation during training.
The training condition evaluation is composed of a loop over the examples and a switch over the condition type (and a few other things). Prior to this change, the example loop was outside the condition switch loop, forcing the algorithm to re-check the condition type (and other things) for each examples. After this change, the condition type is outside of the example loop.

Example of speed-ups:
1. Average speed-up of 3.7% on all benchmark.
2. Speed-up of 10-15% on Adult dataset with GBT.
3. Speed-up of 8% on Adult dataset with RF.
4. Speed-up of 9% on 4M dataset with 200 features with discretized GBT.
5. No speed difference (<1% gain) on 4M dataset with 200 features non-discretized GBT. Note: The absolute gain is the same as 4., but since 5.'s training is longer, the relative gain is insignificant.

PiperOrigin-RevId: 670888905