[YDF] Fix output_logits for slow prediction engine
This unifies the handling of output_logits in the slow prediction engines. This is a breaking change (see below).

If `output_logits` is set on a model, the slow engine will populate the new `logits` field on the prediction proto. Otherwise, this field is not populated. The `distribution` field on the proto is **always** with the probabilities.
The latter is a breaking change / bugfix, since the distribution field was previously populated with the logits in some cases.

PiperOrigin-RevId: 825085209